% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{calc_rpp_bc}
\alias{calc_rpp_bc}
\title{Calculate the proportion of recurrent parent background (RPP) fully recovered
in backcross progenies.}
\usage{
calc_rpp_bc(
  x,
  map_file,
  map_chr = "chr",
  map_pos = "pos",
  map_snp_ids = "snpid",
  rp = NULL,
  rp_num_code = 1,
  na_code = NA,
  weighted = TRUE
)
}
\arguments{
\item{x}{A numeric matrix of marker genotypes for backcross progenies and recurrent
parent. Markers are columns and samples are rows.}

\item{map_file}{A data frame of map file consisting of marker IDs and their
chromosome numbers and positions as columns.}

\item{map_chr}{A character value indicating the column name for chromosome IDs
in \code{map_file}.}

\item{map_pos}{A character value indicating the column name for chromosome
positions in map file.}

\item{map_snp_ids}{A character value indicating the column name for SNP IDs
in \code{map_file}.}

\item{rp}{An integer or character value indicating the row index or sample ID
for the recurrent parent. if `NULL` the sample on the first row of \code{x} is
used as the recurrent parent.}

\item{rp_num_code}{A numeric value indicating the coding for the recurrent parent
marker background in \code{x}.}

\item{na_code}{A value indicating missing data in \code{x}.}

\item{weighted}{A logical value indicating whether RPP values should be weighted
or not.}
}
\value{
A data frame object comprising the RPP of sample IDs.
}
\description{
Calculate the proportion of recurrent parent background (RPP) fully recovered
in backcross progenies.
}
\details{
The weighted RPP is computed as the sum of the product of relative distance
weighting of markers and genotype scores that match the recurrent parent.
Marker weights are obtained as half the normalized ratio of the physical
distance between marker `i` and marker `i+1` to the total distance for all
ordered markers on a chromosome.

The computation excludes heterozygous loci and only considers chromosome regions
fully recovered as recurrent parent genetic background.
}
\examples{
\donttest{
# example code
library(panGenomeBreedr)
# Create a numeric matrix of genotype scores for 10 markers and 5 samples
num_dat <- matrix(c(rep(1, 10), rep(0, 10),
                    1, 1, 0.5, 1, 1, 1, 1, 1, 0, 1,
                    1, 1, 0, 1, 1, 1, 1, 1, 1, 1,
                    1, 1, 0, 1, 1, 1, 1, 1, 1, 0.5 ),
                  byrow = TRUE, ncol = 10)

rownames(num_dat) <- c('rp', 'dp', paste0('bc1_', 1:3))
colnames(num_dat) <- paste0('S1', '_', c(floor(seq(1000, 10000, len = 8)),
                                         15000, 20000))

# Get map file by parsing SNP IDs
map_file <- parse_marker_ns(colnames(num_dat))

# Calculate weighted RPP
rpp <- calc_rpp_bc(x = num_dat,
                      map_file = map_file,
                      map_chr = 'chr',
                      map_pos = 'pos',
                      map_snp_ids = 'snpid',
                      rp = 1,
                      rp_num_code = 1,
                      weighted = TRUE)
}

}
