% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_rpp_exp}
\alias{calc_rpp_exp}
\title{Compute theoretical RPP values for any specified backcross generation.}
\usage{
calc_rpp_exp(bc_gen = 1, rpp2n = FALSE)
}
\arguments{
\item{bc_gen}{A positive integer value indicating the backcross generation.}

\item{rpp2n}{A logical value indicating whether to compute theoretical RPP
values for the F1 all BC generations up to \code{bc_gen}.}
}
\value{
A named vector of theoretical RPP values.
}
\description{
Compute theoretical RPP values for any specified backcross generation.
}
\examples{
# example code
# Calculate theoretical RPP values up to BC5
rpp_values <- calc_rpp_exp(bc_gen = 5, rpp2n = TRUE)

}
