% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{connect_local_db}
\alias{connect_local_db}
\title{Connect to the local offline database}
\usage{
connect_local_db(folder_path, max_memory = "8GB", quiet = FALSE)
}
\arguments{
\item{folder_path}{Character. The absolute or relative path to the directory 
containing the core Parquet files (`genotypes.parquet`, `variants.parquet`, 
`metadata.parquet`, `annotations.parquet`).}

\item{max_memory}{Character. Maximum RAM DuckDB is allowed to utilize. 
Defaults to '"8GB"'. Adjust based on system constraints.}

\item{quiet}{Logical. If \code{TRUE}, suppresses console messages. Defaults to \code{FALSE}.}
}
\value{
A DBI connection object (`con`) linked to the virtual database.
}
\description{
Connect to the local offline database
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder, max_memory = "8GB")

}

}
