% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{create_tabix_bash}
\alias{create_tabix_bash}
\title{Create a tabix bash file to run on HPC cluster}
\usage{
create_tabix_bash(
  slurm_par = c(nodes = 1, ntasks = 1, mem_per_cpu = "1G", time = "00:05:00"),
  cand_gene_id = "Sobic.003G421300",
  tabix_output_path,
  gff_path,
  snpeff_snp = TRUE,
  snpeff_indel = TRUE,
  vcf_path,
  vcf_file_prefix,
  snp_vcf_suffix,
  indel_vcf_suffix,
  bash_out_path = tempdir(),
  filename = "tabix.sh"
)
}
\arguments{
\item{slurm_par}{A named character vector of `length = 4`, indicating SLURM job
parameters for submitting jobs to a computing cluster managed by a SLURM
workload manager. It must be specified in this order: number of compute nodes,
number of tasks, the amount of memory per CPU core, and the maximum runtime
for the job.}

\item{cand_gene_id}{A character value specifying the candidate gene ID.}

\item{tabix_output_path}{A character value indicating the path to directory for
saving tabix results output.}

\item{gff_path}{A character value indicating the path to the GFF file.}

\item{snpeff_snp, snpeff_indel}{A logical value indicating whether to run tabix
command on SNP or InDel snpEff annotated files.}

\item{vcf_path}{A character value indicating the path to directory containing
snpEff annotated SNP and InDel VCF files.}

\item{vcf_file_prefix}{A character avlue indicating the prefix for snpEff
annotated VCF files.}

\item{snp_vcf_suffix, indel_vcf_suffix}{A character value indicating the suffix
for snpEff annotated VCF files for SNPs or InDels.}

\item{bash_out_path}{A character value indicating the path to directory for
saving tabix bash file.}

\item{filename}{A character value indicating the file name of the tabix bash
file.}
}
\value{
A tabix bash file for extracting snpEff annotated variants in SNP or
InDel VCF files for any candidate gene.
}
\description{
Create a tabix bash file to run on HPC cluster
}
\examples{
# example code
library(panGenomeBreedr)

# Create a tabix bash file to extract snpEff annotation for Sobic.003G421300
# Path to GFF file
path1 <- "/pl/active/Morris_CSU/Clara_Cruet"
gff_path <- file.path(path1, 'resequence_pipeline_source_data/genes.gff')

# Path to snpEff annotated vcf file
path2 <- "/pl/active/Morris_CSU/Sorghum_Genetic_Data/Sbicolor_v5.1"
vcf_path <- file.path(path2, 'VCF_update_data_nov/imputed_data_snpeff')

create_tabix_bash(slurm_par = c(nodes = 1,
                                ntasks = 1,
                                mem_per_cpu = '1G',
                                time ='00:05:00'),
                  cand_gene_id = 'Sobic.003G421300',
                  tabix_output_path = "/scratch/alpine/awkena@colostate.edu/ssh_try/",
                  gff_path = gff_path,
                  vcf_path = vcf_path,
                  vcf_file_prefix = "Sorghum_d8.noduplicates.",
                  snp_vcf_suffix = ".snp._markernamesadded_imputed_snpeff",
                  indel_vcf_suffix = ".indel._markernamesadded_imputed_snpeff",
                  bash_out_path = tempdir(),
                  filename = 'tabix_test.sh')

# Path to read bash script
tabix_script <- file.path(tempdir(), 'tabix_test.sh')

# Read the bash file into R
tabix_content <- readLines(tabix_script)
print(tabix_content)

}
