% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{cross_qc_ggplot}
\alias{cross_qc_ggplot}
\title{Create a heatmap to visualize and compare the genetic genetic backgrounds of
genotypes/lines.}
\usage{
cross_qc_ggplot(
  x,
  snp_ids = "snpid",
  geno_ids = "x",
  chr = "chr",
  chr_pos = "pos",
  value = "value",
  parents,
  group_sz = 10L,
  pdf = FALSE,
  filename = "background_heatmap",
  legend_title = "Heatmap_key",
  col_mapping,
  col_labels,
  alpha = 1,
  text_size = 12,
  width = 10.5,
  height = 8.5,
  ...
)
}
\arguments{
\item{x}{A data frame object obtained by melting a numeric matrix of genotype
calls. It must contain the marker IDs, sample IDs, chromosome IDs and positions,
as well as the values for the numeric codes.}

\item{snp_ids}{A character value indicating the column name for marker IDs
in \code{x}.}

\item{geno_ids}{A character value indicating the column name for sample or
genotype IDs in \code{x}.}

\item{chr}{A character value indicating the column name for chromosome IDs
in \code{x}.}

\item{chr_pos}{A character value indicating the column name for chromosome
positions in \code{x}.}

\item{value}{A character value indicating the column name for the values for
the numeric codes in \code{x}.}

\item{parents}{A character vector of length = 2 for the IDs of parents.}

\item{group_sz}{A positive integer value indicating the batch size for progenies
to include in heatmap.}

\item{pdf}{A logical value indicating whether to save plot as a pdf graphic
device when TRUE or output plot in R when FALSE.}

\item{filename}{A character value for path or file name for saving pdf.}

\item{legend_title}{A character value for specifying plot legend title.}

\item{col_mapping}{A character vector of length = 6 for heatmap color mapping.}

\item{col_labels}{A character vector of length = 6 for labels corresponding to
the color mapping.}

\item{alpha}{A numeric value between 0 and 1 for modifying the
opacity of colors.}

\item{text_size}{A numeric value for setting text size.}

\item{width}{A numeric value for the width of pdf device.}

\item{height}{A numeric value for the height of pdf device.}

\item{...}{Other valid arguments that can be passed to ggplot2.}
}
\value{
A graphic object or ggplot.
}
\description{
Create a heatmap to visualize and compare the genetic genetic backgrounds of
genotypes/lines.
}
\examples{
\donttest{
# example code
library(panGenomeBreedr)

# Set path to the directory where your data is located
path1 <-  system.file("extdata", "agriplex_dat.csv",
                      package = "panGenomeBreedr",
                      mustWork = TRUE)

# Import raw Agriplex data file
geno <- read.csv(file = path1, header = TRUE, colClasses = c("character"))

# Get map file by parsing SNP IDs
snps <- colnames(geno)[-c(1:6)] # Get snp ids
map_file <- parse_marker_ns(x = snps, sep = '_', prefix = 'S')

# Process genotype data to re-order SNPs based on chromosome and positions
stg5 <- proc_kasp(x = geno[geno$Batch == 3,], # stg5 NILs
                  kasp_map = map_file,
                  map_snp_id = "snpid",
                  sample_id = "Genotype",
                  marker_start = 7,
                  chr = 'chr',
                  chr_pos = 'pos')

map_file <- stg5$ordered_map # Ordered map
stg5 <- stg5$ordered_geno # Ordered geno

# Convert to numeric format for plotting
num_geno <- kasp_numeric(x = stg5,
                         rp_row = 1, # Recurrent parent row ID
                         dp_row = 3, # Donor parent row ID
                         sep = ' / ',
                         data_type = 'agriplex')

# Melt num_geno into a tidy data frame
df <- gg_dat(num_mat = num_geno,
             map_file = map_file,
             map_pos = 'pos',
             map_chr = 'chr',
             map_snp_ids = 'snpid')

# Create a heatmap that compares the parents to progenies
cross_qc_ggplot(x = df,
                snp_ids = 'snpid',
                geno_ids = 'x',
                chr = 'chr',
                chr_pos = 'pos',
                value = 'value',
                parents = c('BTx623', 'BTx642'),
                group_sz = 5L,
                pdf = FALSE,
                filename = 'background_heatmap',
                legend_title = 'Heatmap_key',
                alpha = 0.8,
                text_size = 14,
                width = 12,
                height = 10)
}

}
