% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{disconnect_local_db}
\alias{disconnect_local_db}
\title{Disconnect the local database session}
\usage{
disconnect_local_db(con, quiet = FALSE)
}
\arguments{
\item{con}{A valid DBI connection object generated by \code{connect_local_db()}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses console messages. Defaults to \code{FALSE}.}
}
\value{
None. Prints a message confirming the disconnection status.
}
\description{
Disconnect the local database session
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder, max_memory = "8GB")

# Disconnect at the end of your session
disconnect_local_db(con)

}

}
