% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{fetch_variants_by_allele_frequency}
\alias{fetch_variants_by_allele_frequency}
\title{Fetch variants by allele frequency}
\usage{
fetch_variants_by_allele_frequency(
  con = NULL,
  min_af = 0,
  max_af = 1,
  chrom = NULL,
  start = NULL,
  end = NULL,
  connect_db_mode = c("local", "online")
)
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{min_af}{Numeric. Minimum alternate allele frequency threshold for filtering.
Must fall between 0 and 1. Defaults to \code{0}.}

\item{max_af}{Numeric. Maximum alternate allele frequency threshold for filtering.
Must fall between 0 and 1. Defaults to \code{1}.}

\item{chrom}{A character value specifying the target chromosome name (e.g., \code{"Chr05"}).}

\item{start}{Integer. Start coordinate for the target window region.}

\item{end}{Integer. End coordinate for the target window region.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A data frame of variants passing the frequency thresholds within the defined region,
  tracking their coordinates, calculated reference allele frequencies (\code{ref_af}),
  and alternate allele frequencies (\code{alt_af}).
}
\description{
This function queries variants within a specified genomic region and filters
them based on a minimum and maximum alternate allele frequency. It can operate
in either 'local' mode (querying a local DuckDB database) or 'online' mode
(fetching data from a remote API endpoint).
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)

# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# Query a specific locus window and filter for common variants (MAF >= 5\%)
local_common_vars <- fetch_variants_by_allele_frequency(con = con_local,
                                min_af = 0.05,
                                max_af = 0.95,
                                chrom = "Chr05",
                                start = 75104537,
                                end = 75106403)
print(head(local_common_vars))

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

\dontrun{
# To query the public online resource instead:
online_common_vars <- fetch_variants_by_allele_frequency(
  min_af = 0.05,
  max_af = 0.95,
  chrom = "Chr05",
  start = 75104537,
  end = 75106403,
  connect_db_mode = 'online'
)
print(head(online_common_vars))
}
}
