% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{geno_error}
\alias{geno_error}
\title{Identify SNP loci with potential genotype call errors.}
\usage{
geno_error(
  x,
  rp_row,
  dp_row,
  sep = ":",
  blank = "NTC",
  uncallable = "Uncallable",
  unused = "?",
  others = c("Missing", "Bad", "Dupe", "Over", "Short")
)
}
\arguments{
\item{x}{A data frame of genotype calls where all columns are SNPs and samples
as rows. Row names are unique sample names.}

\item{rp_row}{An integer indication the row index of the recurrent or Parent 1.}

\item{dp_row}{An integer indicating the row index of the donor or Parent 2.
on failed genotype call.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}
}
\value{
A list object with the following components:
1) column index of loci with genotype call error
2) data frame of loci with genotype call errors if present.
}
\description{
Identify SNP loci with potential genotype call errors.
}
\examples{
# example code
\donttest{
#' library(panGenomeBreedr)
# Reshape KASP data for each master plate for the beta_carotene data
plate_wide <- kasp_reshape_wide(x = beta_carotene,
                                subset = 'MasterPlate',
                                snp_id = 'SNPID',
                                geno_call = 'Call',
                                idvar = "SubjectID",
                                blank = 'NTC')

# Get Master Plate 1
plate1 <- plate_wide$`MWCP-PLATE-1 SE-23-2021_P01`

plate1 <- plate_wide$`SE-24-1088_P01_d1`

# Generate a map for the beta_carotene KASP data
kasp_map <- data.frame(SNPID = unique(beta_carotene$SNPID),
                       SNPID_2 = paste0('snpSB', 1:4, '|', c(800, 803, 804, 805)),
                       chr = c(1:4),
                       pos = c(800, 803, 804, 805))

# Process Plate1 to re-order SNPs based on chrom. and position
proc_plate1 <- proc_kasp(x = plate1,
                         kasp_map = kasp_map,
                         map_snp_id = "SNPID",
                         sample_id = "SubjectID",
                         marker_start = 2,
                         chr = 'chr',
                         chr_pos = 'pos')

# Check for genotype call  error for each SNP
geno_mat <- geno_error(x = proc_plate1$ordered_geno,
                  rp = 1,
                  dp = 2,
                  sep = ':')
}

}
