% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{get_alleles}
\alias{get_alleles}
\title{Get SNP or InDel alleles and possible genotypes from genotype calls in KASP data.}
\usage{
get_alleles(x, sep = ":", data_type = c("kasp", "agriplex"))
}
\arguments{
\item{x}{A character vector of KASP genotype calls in one reaction plate.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{data_type}{A character value indicating the data source; either `kasp` or
`agriplex`.}
}
\value{
A list object with `length = 2` consisting of marker alleles and
possible genotypes in \code{x}.
}
\description{
Get SNP or InDel alleles and possible genotypes from genotype calls in KASP data.
}
\examples{
# example code
\donttest{
# Simulate a typical KASP genotype call
set.seed(123)
x <- sample(c("A:A", "A:-", "-:-", "Uncallable", "?"),
            size = 96,
            replace = TRUE)

# Assign NTC wells
x[c(88, 96)] <- 'NTC'

# Get alleles and expected genotypes
alleles <- get_alleles(x = x, data_type = 'kasp')
}

}
