% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_validation.R
\name{kasp_color}
\alias{kasp_color}
\title{Color-code KASP genotype calls based on LGC genomics colors for HEX and FAM.}
\usage{
kasp_color(
  x,
  subset = "MasterPlate",
  sep = ":",
  geno_call = "Call",
  uncallable = "Uncallable",
  unused = "?",
  blank = "NTC",
  others = c("Missing", "Bad", "Dupe", "Over", "Short"),
  assign_cols = c(FAM = "blue", HEX = "gold", het = "forestgreen")
)
}
\arguments{
\item{x}{A data frame of KASP genotype calls for one or multiple plates.}

\item{subset}{A character value indicating the column name for taking subsets
of \code{x} for processing; default is the `MasterPlate`.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{geno_call}{A character indicating the column name used for genotype
calls in \code{x}.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}

\item{assign_cols}{A named character vector of `length = 3` for assigning colors
to the FAM, HEX and heterozygous genotype groups.}
}
\value{
A list object with subset unit as component data frames.
}
\description{
Color-code KASP genotype calls based on LGC genomics colors for HEX and FAM.
}
\details{
This is an experimental function. The default values of some of the
arguments in the function are based on LGC Genomics conventions including
the color codes for FAM and HEX fluorescence.
}
\examples{
# example code
library(panGenomeBreedr)
\donttest{
dat1 <- kasp_color(x = panGenomeBreedr::kasp_dat,
                   subset = 'MasterPlate',
                   sep = ':',
                   geno_call = 'Call',
                   uncallable = 'Uncallable',
                   unused = '?',
                   blank = 'NTC')
}

}
