% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{kasp_pch}
\alias{kasp_pch}
\title{Generate pch characters for cluster plots of KASP genotype calls.}
\usage{
kasp_pch(
  x,
  sep = ":",
  blank = "NTC",
  uncallable = "Uncallable",
  unused = "?",
  others = c("Missing", "Bad", "Dupe", "Over", "Short")
)
}
\arguments{
\item{x}{A character vector of KASP genotype calls in one resaction plate.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}
}
\value{
A `96 x 1` data frame  of pch values for possible genotypes in
each KASP reaction plate.
}
\description{
Generate pch characters for cluster plots of KASP genotype calls.
}
\examples{
# example code
\donttest{
x <- panGenomeBreedr::kasp_dat$Call[1:96]
geno_pch <- kasp_pch(x = x)
}

}
