% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{kasp_qc_ggplot2}
\alias{kasp_qc_ggplot2}
\title{Make KASP marker genotyping QC plot overlaid with predicitons.}
\usage{
kasp_qc_ggplot2(
  x,
  FAM = "X",
  HEX = "Y",
  geno_call = "Call",
  color = "Color",
  PCH = "pch",
  snp_id = "SNPID",
  blank = "NTC",
  uncallable = "Uncallable",
  unused = "?",
  others = c("Missing", "Bad", "Dupe", "Over", "Short"),
  Group_id = NULL,
  Group_unknown = "?",
  pred_cols = c(Blank = "black", False = "red", True = "blue", Unknown = "yellow2"),
  scale = FALSE,
  pdf = TRUE,
  width = 6,
  height = 6,
  filename = "kasp_qc",
  expand_axis = 0.5,
  legend.pos.x = 0.6,
  legend.pos.y = 0.75,
  legend.box = "horizontal",
  legend.pos = "inside",
  alpha = 0.9,
  text_size = 12,
  ...
)
}
\arguments{
\item{x}{A list object of KASP genotype calls processed by the `kasp_color()`
function.}

\item{FAM}{A character indicating the column name for FAM fluorescence
coordinates in \code{x}.}

\item{HEX}{A character indicating the column name for HEX fluorescence
coordinates in \code{x}.}

\item{geno_call}{A character value indicating the column name for KASP genotype
calls in \code{x}.}

\item{color}{A character value indicating the column name for assigned colors
in \code{x}.}

\item{PCH}{A character value indicating the column name for assigned PCH symbols
in \code{x}.}

\item{snp_id}{A character value indicating the column name for SNP IDs
in \code{x}.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}

\item{Group_id}{A character value for the column ID indicating the predictions
of the positive controls in \code{x}.}

\item{Group_unknown}{A character value representing unknown expected genotype status
for samples, if present. No genotype prediction could be made for such samples.}

\item{pred_cols}{A named character vector of length = 4 of colors to be used for
the prediction legend for positive controls, if present.}

\item{scale}{A logical value indicating whether to scale FAM and HEX axis
to values between 0 and 1.}

\item{pdf}{A logical value indicating whether to save plot as a pdf graphic
device when TRUE or output plot in R when FALSE.}

\item{width}{A numeric value for the width of pdf device.}

\item{height}{A numeric value for the height of pdf device.}

\item{filename}{A character value for path or file name for saving pdf.}

\item{expand_axis}{A numeric value to expand the axes for legend positioning.}

\item{legend.pos.x}{A numeric value representing the x coordinate for legend
placement.}

\item{legend.pos.y}{A numeric value representing the y coordinate for legend
placement.}

\item{legend.box}{A character value of either `horizontal` or `vertical`
legend placement.}

\item{legend.pos}{A character value for the position of legend; the
default value is `inside`.}

\item{alpha}{A numeric value between 0 and 1 for modifying the
opacity of colors.}

\item{text_size}{A numeric value for setting text size.}

\item{...}{Other valid arguments that can be passed to ggplot2.}
}
\value{
A graphic object or plot.
}
\description{
Make KASP marker genotyping QC plot overlaid with predicitons.
}
\examples{
# example code
library(panGenomeBreedr)
\donttest{
# Assign KASP colors to plates
dat1 <- kasp_color(x = panGenomeBreedr::kasp_dat,
                   subset = 'MasterPlate',
                   sep = ':',
                   geno_call = 'Call',
                   uncallable = 'Uncallable',
                   unused = '?',
                   blank = 'NTC')

# KASP QC plot for Plate 12
kasp_qc_ggplot2(x = dat1[12],
                    pdf = FALSE,
                    Group_id = 'Group',
                    Group_unknown = NULL,
                    scale = TRUE,
                    expand_axis = 0.6,
                    alpha = 0.5,
                    legend.pos.x = 0.6,
                    legend.pos.y = 0.8)
}

}
