% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_validation.R
\name{kasp_reshape_wide}
\alias{kasp_reshape_wide}
\title{Reshape KASP data to wide format for same samples genotyped with multiple KASP
markers.}
\usage{
kasp_reshape_wide(
  x,
  subset = "MasterPlate",
  snp_id = "SNPID",
  geno_call = "Call",
  idvar = "SubjectID",
  blank = "NTC"
)
}
\arguments{
\item{x}{A data frame of KASP genotype calls for one or multiple plates.}

\item{subset}{A character value indicating the column name for master plate
having the same samples in \code{x}.}

\item{snp_id}{A character value indicating the column name for SNP IDs
in \code{x}.}

\item{geno_call}{A character value indicating the column name for KASP genotype
calls in \code{x}.}

\item{idvar}{A character value indicating the column name for unique subject
IDs of samples in \code{x}.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}
}
\value{
A list object with reshaped data for master plates that have the same
subject IDs as the components. The components are data frames with Column 1 as
the subject IDs and the rest of the columns as the number of KASP markers assayed
for each master plate.
}
\description{
Reshape KASP data to wide format for same samples genotyped with multiple KASP
markers.
}
\examples{
\donttest{
# example code
library(panGenomeBreedr)
dat1 <- panGenomeBreedr::beta_carotene
plate_wide <- kasp_reshape_wide(x = dat1,
                                subset = 'MasterPlate',
                                snp_id = 'SNPID',
                                geno_call = 'Call',
                                idvar = "SubjectID",
                                blank = 'NTC')
}

}
