% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{list_tables}
\alias{list_tables}
\title{List all tables in the database}
\usage{
list_tables(con = NULL, connect_db_mode = c("local", "online"))
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A character vector of table names in the database.
}
\description{
This function lists all available tables from either a local database connection
or the online database, depending on the selected mode.
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)

# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# List tables from the local connection (default mode is 'local')
local_tables <- list_tables(con = con_local)
print(local_tables)

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

\dontrun{
# To query the public online resource instead:
online_tables <- list_tables(connect_db_mode = 'online')
print(online_tables)
}

}
