% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{nsamples_plate}
\alias{nsamples_plate}
\title{Get a summary of the number of samples per 96-well plate in a multi-plate KASP
assay.}
\usage{
nsamples_plate(
  x,
  subset = "plates",
  snp_id = "SNPID",
  plate_id = "MasterPlate"
)
}
\arguments{
\item{x}{A data frame of KASP genotype calls for one or multiple plates.}

\item{subset}{A character value indicating the column name for taking subsets
of \code{x} for processing; default is the `plates`.}

\item{snp_id}{A character value indicating the column name for SNP IDs
in \code{x}.}

\item{plate_id}{A character value indicating the column name for master plate
having the same samples in \code{x}.}
}
\value{
A list object with plates and a summary of number of samples
per plate as components.
}
\description{
Get a summary of the number of samples per 96-well plate in a multi-plate KASP
assay.
}
\examples{
\donttest{
# example code
library(panGenomeBreedr)
dat1 <- panGenomeBreedr::beta_carotene
dat1 <- nsamples_plate(x = dat1,
                     subset = 'plates',
                     snp_id = 'SNPID',
                     plate_id = 'MasterPlate'
                    )$summ
}

}
