% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_markers}
\alias{order_markers}
\title{Order marker IDs based on their chromosome numbers and positions in ascending
order.}
\usage{
order_markers(x, chr_col = "chr", pos_col = "pos")
}
\arguments{
\item{x}{A data frame object for a map file or marker data file.}

\item{chr_col}{A character value indicating the column name for chromosome IDs
in \code{x}.}

\item{pos_col}{A character value indicating the column name for chromosome
positions in \code{x}.}
}
\value{
A data frame object of the same dimension as \code{x}.
}
\description{
Order marker IDs based on their chromosome numbers and positions in ascending
order.
}
\examples{
# example code
map_file <- data.frame(snpid = paste0('S', rep(1:2, 5), '_', 1001:1005),
                       chr = rep(1:2, 5),
                       pos = rep(1001:1005, 2))

# Order map file
map_file <- order_markers(x = map_file,
                          chr_col = 'chr',
                          pos_col = 'pos')

}
