% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{parent_poly}
\alias{parent_poly}
\title{Select polymorphic loci between two parents in a marker panel.}
\usage{
parent_poly(x, rp_row, dp_row, sep = ":", na_code = NA)
}
\arguments{
\item{x}{A data matrix or frame with markers as columns and samples as rows.}

\item{rp_row, dp_row}{An integer or character value indicating the row index
or name of Parent 1 and 2.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{na_code}{A value indicating missing data in \code{x}.}
}
\value{
A data matrix or frame object of polymorphic loci between two parental
lines.
}
\description{
Select polymorphic loci between two parents in a marker panel.
}
\examples{
# example code

# Marker data
dat <- data.frame(snp1 = c('C:C', 'A:A', 'C:A', 'C:A'),
                  snp2 = c('C:C', 'C:C', 'C:C', 'C:C'),
                  snp3 = c('T:T', 'C:C', 'C:T', 'C:T'),
                  snp4 = c('G:G', '-:-', 'G:-', 'G:G'),
                  snp5 = c('T:T', 'T:T', 'T:A', 'T:A'),
                  row.names = c('rp', 'dp', 'art_het1', 'art_het2'))

# Find polymorphic loci
poly_loci <- parent_poly(x = dat,
                         rp_row = 1,
                         dp_row = 2,
                         sep = ':')

}
