% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_count_variant_types}
\alias{pg_count_variant_types}
\title{Count the distribution of variant types in the database (online)}
\usage{
pg_count_variant_types(variants_table = "variants")
}
\arguments{
\item{variants_table}{Character. The name of the table containing variant
metadata. Defaults to "variants".}
}
\value{
A data frame with two columns: \code{variant_type} and \code{n}.
}
\description{
Count the distribution of variant types in the database (online)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Count the distribution of variant types in the online database
pg_count_variant_types_result <- pg_count_variant_types()
print(pg_count_variant_types_result)
}


}
