% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{pg_list_table_columns}
\alias{pg_list_table_columns}
\title{List column names and types for a specified table (online)}
\usage{
pg_list_table_columns(
  table_name = c("variants", "annotations", "genotypes", "metadata")
)
}
\arguments{
\item{table_name}{A character value specifying the name of the table.
Defaults to "variants". Available options are "variants", "annotations",
"genotypes", or "metadata".}
}
\value{
A data frame containing column metadata.
}
\description{
This function connects to the public panGenomeBreedr API and retrieves
metadata about the columns in a specified table.
}
