% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_variant_impact_summary}
\alias{pg_variant_impact_summary}
\title{Get summary of database variant impacts by chromosome (online)}
\usage{
pg_variant_impact_summary()
}
\value{
A data frame in wide format where each row is a chromosome and
columns represent the counts for each impact category.
}
\description{
Get summary of database variant impacts by chromosome (online)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Get variant impact summary from the online database
pg_variant_impact_summary_result <- pg_variant_impact_summary()
print(pg_variant_impact_summary_result)
}

}
