% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcil_script.R
\name{plot_pcil_positive}
\alias{plot_pcil_positive}
\title{Plot PCIL Positive Introgressions}
\usage{
plot_pcil_positive(pcil_positive_result)
}
\arguments{
\item{pcil_positive_result}{A list object returned by `fetch_pcil_positive`,
containing `pcil_positive` and `regions` data frames.}
}
\value{
A list of ggplot objects, where each plot corresponds to a unique
  target region and visualizes the introgression segments of the positive PCILs.
}
\description{
Visualizes the genomic regions of Pangenome-Characterized Introgression Lines (PCIL)
that were identified as positive for a given target. It generates a plot for each
target region, showing the introgression blocks for all positive lines.
}
\examples{
\donttest{
library(panGenomeBreedr)

# 1. Connect to the package's bundled example database and fetch data
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)
con <- connect_local_db(folder_path = my_db_folder)
pcil_data <- fetch_pcil_data(con = con, connect_db_mode = "local")
selection <- c("INDEL_Chr03_79037889", "SNP_Chr03_79037855")

variant_geno_sel <- fetch_genotypes_by_id(
  con = con,
  variant_ids = selection,
  connect_db_mode = "local"
)

# 2. Fetch relevant families
fam_results <- fetch_pcil_families_by_variant(
  con = con,
  selection = selection,
  pcil_data = pcil_data,
  connect_db_mode = "local"
)

# 3. Select positive lines
pcil_pos_pcv <- fetch_pcil_positive(
  pcil_data = pcil_data,
  variants_select_geno = variant_geno_sel,
  type = "position",
  sel = 15,
  available_ids = fam_results$pcil_summary[, c("sample_id", "selection")],
  result_pcil_families = fam_results,
  window = 0
)

# 4. Generate and display the plot
positive_plots <- plot_pcil_positive(pcil_positive_result = pcil_pos_pcv)

if (length(positive_plots) > 0) {
  print(positive_plots[[1]])
}

disconnect_local_db(con)
}
}
