% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{plot_plate}
\alias{plot_plate}
\title{Plot kasp genotyping plate layout.}
\usage{
plot_plate(
  x,
  well = "MasterWell",
  color = "Color",
  geno_call = "Call",
  snp_id = "SNPID",
  pdf = TRUE,
  width = 8,
  height = 5,
  filename = "plate_layout",
  text_size = 12,
  ...
)
}
\arguments{
\item{x}{A list object of KASP genotype calls processed by the `kasp_color()`
function.}

\item{well}{A character value representing the column name for genotyping
plate wells.}

\item{color}{A character value indicating the column name of assigned colors
in \code{x}.}

\item{geno_call}{A character indicating the column name used for genotype}

\item{snp_id}{A character value indicating the column name for SNP IDs
in \code{x}.}

\item{pdf}{A logical value indicating whether to save plot as a pdf graphic}

\item{width}{A numeric value for the width of pdf device.}

\item{height}{A numeric value for the height of pdf device.}

\item{filename}{A character value for path or file name for saving pdf.}

\item{text_size}{A numeric value for text size in plot output.}

\item{...}{Other valid arguments that can be passed to ggplot2.}
}
\value{
A ggplot graphical output of plate layout.
}
\description{
Plot kasp genotyping plate layout.
}
\examples{
# example code
library(panGenomeBreedr)
\donttest{
# Assign KASP colors to plates
dat1 <- kasp_color(x = panGenomeBreedr::kasp_dat,
                   subset = 'MasterPlate',
                   sep = ':',
                   geno_call = 'Call',
                   uncallable = 'Uncallable',
                   unused = '?',
                   blank = 'NTC')
# Plot Plate 12 to see sample arrangement
plot_plate(dat1[12], pdf = FALSE)
}

}
