% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{proc_kasp}
\alias{proc_kasp}
\title{Process reshaped KASP genotype data for heatmap plotting}
\usage{
proc_kasp(
  x,
  sample_id = "SubjectID",
  marker_start = 2,
  kasp_map,
  map_snp_id,
  chr = "chr",
  chr_pos = "pos"
)
}
\arguments{
\item{x}{A data frame of genotype calls.}

\item{sample_id}{A string representing the column name of unique sample IDs.}

\item{marker_start}{An integer indicating the column index for the start of SNP calls.}

\item{kasp_map}{A data frame consisting of marker chromosome number and positions.}

\item{map_snp_id}{A character value indicating the column name for SNP IDs
in \code{kasp_map}.}

\item{chr}{A character value for the column name for the chromosome number in
\code{kasp_map}.}

\item{chr_pos}{A character value for the column name for the chromosome positions
in \code{kasp_map}.}
}
\value{
A list consisting of re-ordered SNPs and oredered map based on chromosome
number and chromosome positions.
}
\description{
Process reshaped KASP genotype data for heatmap plotting
}
\details{
This function is experimental and should only be used with reshaped KASP data.
The data processing involves the following steps:
1. Transposing the data
2. Matching marker names in KASP map file with names in KASP genotype file
3. Column bind matched map for markers to the transposed data
4. Sort data in ascending order of chromosome numbers
5. Split data by chromosomes
6. Sort data in ascending order of physical positions per chromosome
7. Row bind sorted data together
8. Re-transpose data to make markers columns and samples as rows for plotting.
}
\examples{
# example code
\donttest{
#' library(panGenomeBreedr)
# Reshape KASP data for each master plate for the beta_carotene data
plate_wide <- kasp_reshape_wide(x = beta_carotene,
                                subset = 'MasterPlate',
                                snp_id = 'SNPID',
                                geno_call = 'Call',
                                idvar = "SubjectID",
                                blank = 'NTC')

# Get Master Plate 1
plate1 <- plate_wide$`MWCP-PLATE-1 SE-23-2021_P01`

plate1 <- plate_wide$`SE-24-1088_P01_d1`

# Generate a map for the beta_carotene KASP data
kasp_map <- data.frame(SNPID = unique(beta_carotene$SNPID),
                       SNPID_2 = paste0('snpSB', 1:4, '|', c(800, 803, 804, 805)),
                       chr = c(1:4),
                       pos = c(800, 803, 804, 805))

# Process Plate1 to re-order SNPs based on chrom. and position
proc_plate1 <- proc_kasp(x = plate1,
                         kasp_map = kasp_map,
                         map_snp_id = "SNPID",
                         sample_id = "SubjectID",
                         marker_start = 2,
                         chr = 'chr',
                         chr_pos = 'pos')
}
}
