% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{query_by_impact}
\alias{query_by_impact}
\title{Extract variants based on functional mutation impact (local)}
\usage{
query_by_impact(
  con,
  impact_level = c("HIGH", "MODERATE", "LOW", "MODIFIER"),
  chrom = NULL,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{impact_level}{A character vector specifying the variant impact types to retain.
Allowed classifications are \code{"HIGH"}, \code{"MODERATE"}, \code{"LOW"}, and \code{"MODIFIER"}.}

\item{chrom}{A character value specifying the chromosome name (e.g., \code{"Chr05"}).}

\item{start}{Integer. Start coordinate for the genomic region.}

\item{end}{Integer. End coordinate for the genomic region.}
}
\value{
A data frame of variants showing their functional impact details
  and positional coordinates.
}
\description{
Extract variants based on functional mutation impact (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Extract high-impact variant lines inside a targeted coordinates locus window
high_impact_vars <- query_by_impact(con = con,
                                    impact_level = "HIGH",
                                    chrom = "Chr05",
                                    start = 75104537,
                                    end = 75106403)

# Disconnect at the end of your session
disconnect_local_db(con)
}

}
