% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{read_kasp_csv}
\alias{read_kasp_csv}
\title{Read raw KASP results file (csv format) with one or multiple plates.}
\usage{
read_kasp_csv(
  file,
  row_tags = c("Statistics", "DNA", "SNPs", "Scaling", "Data"),
  spacing = 2L,
  data_type = c("raw", "polished")
)
}
\arguments{
\item{file}{A character value indicating the file name or path for the KASP
results file (csv format).}

\item{row_tags}{A character vector for the ordered row tags for the
components of the data in \code{file}.}

\item{spacing}{An integer value for specifying the number of empty rows between
data components in \code{file}.}

\item{data_type}{A character value indicating the file type to import;
currently supports either `raw` or `polished` KASP results file.}
}
\value{
A list object of KASP results file ka genotyping calls with FAM and HEX coordinates.
}
\description{
Read raw KASP results file (csv format) with one or multiple plates.
}
