% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{scale_axis}
\alias{scale_axis}
\title{Normalize FAM and HEX fluorescence values between 0 and 1}
\usage{
scale_axis(x)
}
\arguments{
\item{x}{A numeric vector of FAM or HEX fluorescence values}
}
\value{
A numeric vector of normalized values between 0 and 1

A numeric vector of normalized values between 0 and 1.
}
\description{
Normalize FAM and HEX fluorescence values between 0 and 1
}
