% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{connect_local_db}
\alias{connect_local_db}
\title{Connect to the local offline database}
\usage{
connect_local_db(folder_path, max_memory = "8GB", quiet = FALSE)
}
\arguments{
\item{folder_path}{Character. The absolute or relative path to the directory
<<<<<<< HEAD
containing the core database files (`genotypes_partitioned/`, `variants.parquet`,
`metadata.parquet`, `annotations.parquet`) and optionally the `pcil/` extension directory.}
=======
containing the core Parquet files (`genotypes.parquet`, `variants.parquet`,
`metadata.parquet`, `annotations.parquet`).}
>>>>>>> main

\item{max_memory}{Character. Maximum RAM DuckDB is allowed to utilize.
Defaults to '"8GB"'. Adjust based on system constraints.}

\item{quiet}{Logical. If \code{TRUE}, suppresses console messages. Defaults to \code{FALSE}.}
}
\value{
A DBI connection object (`con`) linked to the virtual database.
}
\description{
Establishes a virtual connection to the offline database engine using DuckDB. 
It mounts the core pangenome-scale Parquet files .
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)
my_db_folder <- system.file("extdata", "pangenome_scale_db", 
                           package = "panGenomeBreedr", 
                           mustWork = TRUE)


# Establish the connection
con <- connect_local_db(folder_path = my_db_folder, max_memory = "8GB")

}

}
