% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{count_variant_types}
\alias{count_variant_types}
\title{Count the distribution of variant types in the database}
\usage{
count_variant_types(
  con = NULL,
  variants_table = "variants",
  connect_db_mode = c("local", "online")
)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.
Required only when `connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{variants_table}{Character. The name of the table containing variant
metadata. Defaults to "variants".}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A data frame containing two columns: \code{variant_type} (the structural
  classification group name) and \code{n} (the total count of variants in that group).
}
\description{
This function counts the distribution of variant types (e.g., SNP, INDEL)
from either a local database connection or the online database.
}
\examples{
\donttest{
# --- Online Mode ---
# Load the package
library(panGenomeBreedr)

# Count variant types from the online database
online_counts <- count_variant_types(connect_db_mode = 'online')
print(online_counts)

# --- Offline Mode ---
# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db", 
                           package = "panGenomeBreedr", 
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# Get the breakdown counts of SNPs vs INDELs
local_counts <- count_variant_types(con = con_local)
# print(local_counts)

# Disconnnect at the end of the session.
disconnect_local_db(con_local)
}

}
