% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{count_variant_types}
\alias{count_variant_types}
\title{Count the distribution of variant types in the database (local)}
\usage{
count_variant_types(con, variants_table = "variants")
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{variants_table}{Character. The name of the table containing variant
metadata. Defaults to "variants".}
}
\value{
A data frame containing two columns: \code{variant_type} (the structural
  classification group name) and \code{n} (the total count of variants in that group).
}
\description{
Count the distribution of variant types in the database (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Get the breakdown counts of SNPs vs INDELs
count_variant_types_result <- count_variant_types(con)
print(count_variant_types_result)

# Disconnnect at the end of the session.
disconnect_local_db(con)
}

}
