% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{disconnect_local_db}
\alias{disconnect_local_db}
\title{Disconnect from the Local Database}
\usage{
disconnect_local_db(con, quiet = FALSE)
}
\arguments{
\item{con}{A valid DBI connection object generated by \code{connect_local_db()}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses console messages. Defaults to \code{FALSE}.}
}
\value{
None. Prints a message confirming the disconnection status.
}
\description{
Safely closes the DuckDB connection and clears the virtual Parquet views from memory.
This is the companion function to \code{connect_local_db()} and should be
executed at the end of an analytical session to free up system resources.
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"
con <- connect_local_db(folder_path = my_db_folder)

# Safely close the connection when finished
disconnect_local_db(con)
}

}
