% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{extract_variant}
\alias{extract_variant}
\title{Extract putative causal variants within a candidate gene from a tabix-indexed
snpEff annotated VCF file.}
\usage{
extract_variant(
  cand_gene_id,
  gff_path,
  vcf_dir,
  vcf_file,
  output_path = tempdir(),
  outfile_suffix = "variants"
)
}
\arguments{
\item{cand_gene_id}{A character value specifying the candidate gene ID.}

\item{gff_path}{A character value indicating the path to the GFF file,
including the complete file name.}

\item{vcf_dir}{A character value indicating the path to directory containing
snpEff annotated VCF files.}

\item{vcf_file}{A character value indicating the file name for snpEff
annotated VCF file including the .vcf.gz extension.}

\item{output_path}{A character value indicating the path to directory for
saving extracted variants.}

\item{outfile_suffix}{A character value indicating the file name to be used for
saving extracted variants.}
}
\value{
A VCF file containing extracted variants for \code{cand_gene_id}.
}
\description{
Extract putative causal variants within a candidate gene from a tabix-indexed
snpEff annotated VCF file.
}
\details{
This wrapper function operates on tabix-indexed snpEff annotated VCF files.
However, if a tabix-indexed VCF file is not available, it can create one from
the inputted VCF file.

The file names of snpEff annotated VCF files are expected to consist of three
components: a common prefix, chromosome tag and a common suffix.
}
\examples{
# example code
library(panGenomeBreedr)

# Read the bash file into R
tabix_content <- readLines(tabix_script)
print(tabix_content)

}
