% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{fetch_accession_metadata}
\alias{fetch_accession_metadata}
\title{Get sample metadata from the database}
\usage{
fetch_accession_metadata(
  con = NULL,
  query_col = NULL,
  query_value = NULL,
  connect_db_mode = c("local", "online")
)
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{query_col}{Character. Optional metadata column name to filter by (e.g., `"countryorigin"`).
If `NULL`, all metadata records are returned.}

\item{query_value}{Character or Numeric. The value to match within `query_col`.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A data frame containing the sample metadata records.
}
\description{
Retrieves sample metadata for the genotypes table from the database. It can operate in 'local' mode
to query a local database connection, or 'online' mode to fetch data from
the API. The function can return all metadata or filter by a specific column-value pair.
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)

# --- Online Mode ---
# Retrieve all sample metadata from the online database
all_metadata_online <- fetch_accession_metadata(connect_db_mode = 'online')
print(all_metadata_online[1:3, 1:5])

# Retrieve metadata for samples from Ghana from the online database
ghana_metadata_online <- fetch_accession_metadata(
  query_col = "countryorigin",
  query_value = "Ghana",
  connect_db_mode = 'online'
)
print(ghana_metadata_online[1:3, 1:5])

# --- Offline Mode ---
# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db", 
                           package = "panGenomeBreedr", 
                           mustWork = TRUE)

# Establish a virtual connection
con_local <- connect_local_db(folder_path = my_db_folder)

# Fetch metadata for all sorghum accessions originating from Ethiopia
ethiopia_metadata_local <- fetch_accession_metadata(
  con = con_local,
  query_col = "countryorigin",
  query_value = "Ethiopia"
)
# print(ethiopia_metadata_local[1:3, 1:5])

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

}
