% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{fetch_variants_by_impact}
\alias{fetch_variants_by_impact}
\title{Extract variants based on functional mutation impact}
\usage{
fetch_variants_by_impact(
  con = NULL,
  impact_level = c("HIGH", "MODERATE", "LOW", "MODIFIER"),
  chrom = NULL,
  start = NULL,
  end = NULL,
  connect_db_mode = c("local", "online")
)
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{impact_level}{A character vector specifying the variant impact types to retain.
Allowed classifications are `"HIGH"`, `"MODERATE"`, `"LOW"`, and `"MODIFIER"`.}

\item{chrom}{A character value specifying the chromosome name (e.g., `"Chr05"`).}

\item{start}{Integer. Start coordinate for the genomic region.}

\item{end}{Integer. End coordinate for the genomic region.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A data frame of variants showing their functional impact details
  and positional coordinates.
}
\description{
This function extracts variants based on their predicted functional impact
(e.g., HIGH, MODERATE, LOW, MODIFIER) within a specified genomic region.
It can operate in either 'local' mode (querying a local DuckDB database)
or 'online' mode (fetching data from a remote API endpoint).
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)

# --- Online Mode ---
# Query for high-impact variants in a specific genomic region from the online database
online_high_impact_vars <- fetch_variants_by_impact(
  impact_level = "HIGH",
  chrom = "Chr05",
  start = 75104537,
  end = 75106403,
  connect_db_mode = 'online'
)
print(head(online_high_impact_vars))

# --- Offline Mode ---
# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# Extract high-impact variant lines inside a targeted coordinates locus window from local DB
local_high_impact_vars <- fetch_variants_by_impact(
  con = con_local,
  impact_level = "HIGH",
  chrom = "Chr05",
  start = 75104537,
  end = 75106403,
  connect_db_mode = 'local'
)
# print(head(local_high_impact_vars))

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

}
