% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{gene_coord_gff}
\alias{gene_coord_gff}
\title{Get the genomic range of a candidate gene from a GFF file}
\usage{
gene_coord_gff(gene_name, gff_path)
}
\arguments{
\item{gene_name}{A character value indicating the Sobic ID of the candidate gene.}

\item{gff_path}{A character value specifying the path to the GFF3 file.
URL paths and GZ-compressed files are supported.}
}
\value{
A list object of three components indicating the chromosome, start
and end coordinates of candidate gene.
}
\description{
This function parses a GFF3 file (local or remote) to find the
genomic coordinates (chromosome, start, end) for a specified gene ID.
}
\examples{
\donttest{
# example code

# Define path to a remote sorghum GFF3 file (v5.1)
gff_path <- "https://raw.githubusercontent.com/awkena/panGB/main/Sbicolor_730_v5.1.gene.gff3.gz"

# Retrieve coordinates for a candidate gene
gene_coord_gff(gene_name = "Sobic.005G213600",
               gff_path = gff_path)

}
}
