% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{geno_error}
\alias{geno_error}
\title{Identify SNP loci with potential genotype call errors.}
\usage{
geno_error(
  x,
  rp_row,
  dp_row,
  sep = ":",
  blank = "NTC",
  uncallable = "Uncallable",
  unused = "?",
  others = c("Missing", "Bad", "Dupe", "Over", "Short"),
  data_type = c("kasp", "agriplex")
)
}
\arguments{
\item{x}{A data frame of genotype calls where all columns are SNPs and samples
as rows. Row names are unique sample names.}

\item{rp_row}{An integer or character value indication the row index or name
of the recurrent or Parent 1.}

\item{dp_row}{An integer or character value indicating the row index or name
of the donor or Parent 2.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}

\item{data_type}{A character value indicating the data source; either `kasp` or
`agriplex`.}
}
\value{
A list object with the following components:
1) data frame of loci with genotype call errors if present.
2) data frame of loci with genotype no errors.
}
\description{
Identify SNP loci with potential genotype call errors.
}
\examples{
# example code
\donttest{
library(panGenomeBreedr)
# Reshape KASP data for each master plate for the beta_carotene data
dat1 <- panGenomeBreedr::beta_carotene
plate_wide <- kasp_reshape_wide(x = dat1,
                                subset = 'MasterPlate',
                                snp_id = 'SNPID',
                                geno_call = 'Call',
                                idvar = "SubjectID",
                                blank = 'NTC')

# Get Master Plate 1
plate1 <- plate_wide$`SE-24-1088_P01_d1`

# Check for genotype call  error for each SNP
geno_mat <- geno_error(x = plate1[,-1],
                  rp_row = 1,
                  dp_row = 7,
                  sep = ':',
                  data_type = 'kasp')
}

}
