% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the current API endpoint URL}
\usage{
get_api_url()
}
\value{
The character string of the active API URL.
}
\description{
Retrieves the active API endpoint. If no custom endpoint has been set
by the user, it safely defaults to the official Sorghum database API.
}
