% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the base URL for the panGenomeBreedr API}
\usage{
get_api_url()
}
\value{
A character string of the API base URL.
}
\description{
Get the base URL for the panGenomeBreedr API
}
\keyword{internal}
