% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{hotspot_overlay_plot}
\alias{hotspot_overlay_plot}
\title{Generate a combined gene model and variant hotspot overlay plot (local)}
\usage{
hotspot_overlay_plot(
  gene_name,
  gff_path,
  connect_db_mode = c("online", "local"),
  local_db_path = NULL,
  selected_variants = NULL,
  text_sz = 2.5
)
}
\arguments{
\item{gene_name}{A character string indicating the Sobic ID of the candidate gene.}

\item{gff_path}{A character string specifying the path to the GFF3 file.}

\item{connect_db_mode}{A character string specifying the database connection mode:
either "online" or "local".}

\item{local_db_path}{A character string specifying the local folder path to the database
if running in "local" mode.}

\item{selected_variants}{An optional character vector of specific `variant_id`s to
highlight on the plot with vertical lines, enlarged points, and labels.}

\item{text_sz}{A numeric value for specifying text size for selected variants.}
}
\value{
A \code{patchwork} object containing the combined, aligned plots.
}
\description{
Generates a vertically stacked visualization aligning a structural
gene model with its corresponding regional variant hotspots.
}
\details{
This wrapper function integrates genomic annotations and variant genotypes.
It natively queries the pangenome-scale variant relational database. Whether operating
in 'online' mode or 'local' mode, it leverages the underlying Parquet and DuckDB
architectures optimized for sorghum to efficiently retrieve variants
and their snpEff annotations before plotting.
}
\examples{
\donttest{
library(panGenomeBreedr)
gff_path <- "https://raw.githubusercontent.com/awkena/panGB/main/Sbicolor_730_v5.1.gene.gff3.gz"
hotspot_overlay_plot(gene_name = "Sobic.005G213600",
                     gff_path = gff_path,
                     connect_db_mode = "online",
                     selected_variants = c("INDEL_Chr05_75104881", "INDEL_Chr05_75105587"))
}
}
