% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{kasp_color}
\alias{kasp_color}
\title{Color-code KASP genotype calls based on LGC genomics colors.}
\usage{
kasp_color(
  x,
  subset = "MasterPlate",
  sep = ":",
  geno_call = "Call",
  uncallable = "Uncallable",
  unused = "?",
  blank = "NTC"
)
}
\arguments{
\item{x}{A data frame of KASP genotype calls.}

\item{subset}{A character value indicating the column name for taking subsets
of \code{x} for processing; default is the `MasterPlate`.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{geno_call}{A character indicating the column name used for genotype
calls in \code{x}.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}
}
\value{
A list object with subset unit as component data frames.
}
\description{
Color-code KASP genotype calls based on LGC genomics colors.
}
\details{
This is an experimental function. The default values of some of the
arguments in the function are based on LGC Genomics conventions including
the color codes for FAM and HEX fluorescence.
}
\examples{
# example code
library(panGenomeBreedr)
\donttest{
dat1 <- kasp_color(x = panGenomeBreedr::kasp_dat,
                   subset = 'MasterPlate',
                   sep = ':',
                   geno_call = 'Call',
                   uncallable = 'Uncallable',
                   unused = '?',
                   blank = 'NTC')
}

}
