% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{kasp_marker_design}
\alias{kasp_marker_design}
\title{Design KASP markers based on causal variants.}
\usage{
kasp_marker_design(
  vcf_file,
  marker_ID,
  chr = NULL,
  genome_file,
  plot_draw = TRUE,
  plot_file = tempdir(),
  region_name = "loc_1",
  vcf_geno_code = c("1|1", "0|1", "0|0", ".|."),
  maf = 0.05
)
}
\arguments{
\item{vcf_file}{Path to the vcf file containing identified variants from
re-sequence data.}

\item{marker_ID}{Designated name of variant for marker design. Name must be
contained in `vcf_file`.}

\item{chr}{A character value representation the chromosome description in the
`genome_file`. Providing this helps to save memory in R.}

\item{genome_file}{Path to reference genome file in fasta format, either
compressed (.gz) or decompressed.}

\item{plot_draw}{A logical value indicating whether to plot 100 bp upstream
and downstream KASP sequence alignment to reference genome.}

\item{plot_file}{Path to save the sequence alignment if `plot_draw = TRUE`.}

\item{region_name}{A n optional character value for assigned region name.}

\item{vcf_geno_code}{A character vector for genotype coding for samples
in imported `vcf_file`. The order of the genotype codes in this vector matters:
`c('homo_alt_allele', 'heterozygous', 'homo_ref_allele', 'missing_genotype_call')`.}

\item{maf}{A numeric value between 0 and 1 representing the minor allele
frequency for variant subset.}
}
\value{
A data frame containing all information required for KASP marker
design and a DNA sequence alignment to the reference genome.
}
\description{
Design KASP markers based on causal variants.
}
\details{
This function provides the intertek sequence to be used for marker development
for the selected casual variants. It provides all the information of allele and
location to fill Intertek form.It needs a vcf file of variants calls, and a genome
sequence of the target crop in fasta format.
}
\examples{
\donttest{
# Example to design a KASP marker on a substitution variant
path <- tempdir()
path1 <- "https://raw.githubusercontent.com/awkena/panGB/main/Chr02.fa.gz"
path2 <-  system.file("extdata", "Sobic.002G302700_SNP_snpeff.vcf",
                     package = "panGenomeBreedr",
                     mustWork = TRUE)
ma1 <- kasp_marker_design(vcf_file = path2,
                          genome_file = path1,
                          marker_ID = "SNP_Chr02_69200443",
                          chr = "Chr02",
                          plot_draw = TRUE,
                          plot_file = path,
                          region_name = "ma1",
                          maf = 0.05)

# View marker alignment output from temp folder
path3 <- file.path(path, list.files(path = path, "alignment_"))
# system(paste0('open "', path3, '"'))

on.exit(unlink(path))
}

}
