% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{kasp_numeric}
\alias{kasp_numeric}
\title{Convert processed KASP data to numeric genotypes}
\usage{
kasp_numeric(
  x,
  rp_row,
  dp_row,
  sep = ":",
  blank = "NTC",
  uncallable = "Uncallable",
  unused = "?",
  others = c("Missing", "Bad", "Dupe", "Over", "Short"),
  data_type = c("kasp", "agriplex")
)
}
\arguments{
\item{x}{A data frame of n rows of genotypes and p rows of SNPs; output from
`proc_kasp()` function.}

\item{rp_row}{An integer or character value indicating the row index or name
of the recurrent or Parent 1.}

\item{dp_row}{An integer or character value indicating the row index or name
of donor or Parent 2.}

\item{sep}{A character used as separator for genotype calls, default is a
colon.}

\item{blank}{A character value indicating `No Template Controls (NTC)`
genotype calls.}

\item{uncallable}{A character indicating `Uncallable` genotype calls, if present.}

\item{unused}{A character indicating `?` genotype calls, if present.}

\item{others}{A character vector indicating other non-genotype calls in KASP
genotype calls, if present. These may include `'Missing', 'Bad', 'Dupe'`,
`'Over', 'Short'`.}

\item{data_type}{A character value indicating the data source; either `kasp` or
`agriplex`.}
}
\value{
A data frame of numeric codes for KASP genotype calls.
}
\description{
Convert processed KASP data to numeric genotypes
}
\details{
Re-coded as 1 if homozygous for Parent 1 allele; 0 if homozygous for Parent
2 allele, and 0.5 if heterozygous. If any of the parents SNP call is missing,
it's coded as as NA. If the Parents 1 and 2 genotypes are the same for any snp,
it's coded as monomorphic. Loci with progeny genotype calls different from RP
and DP are coded as -2.
}
\examples{
# example code
\donttest{
#' library(panGenomeBreedr)
# Reshape KASP data for each master plate for the beta_carotene data
plate_wide <- kasp_reshape_wide(x = beta_carotene,
                                subset = 'MasterPlate',
                                snp_id = 'SNPID',
                                geno_call = 'Call',
                                idvar = "SubjectID",
                                blank = 'NTC')

# Get Master Plate 1
plate1 <- plate_wide$`SE-24-1088_P01_d1`

# Generate a map for the beta_carotene KASP data
kasp_map <- data.frame(SNPID = unique(beta_carotene$SNPID),
                       SNPID_2 = paste0('snpSB', 1:4, '|', c(800, 803, 804, 805)),
                       chr = c(1:4),
                       pos = c(800, 803, 804, 805))

# Process Plate1 to re-order SNPs based on chrom. and position
proc_plate1 <- proc_kasp(x = plate1,
                         kasp_map = kasp_map,
                         map_snp_id = "SNPID",
                         sample_id = "SubjectID",
                         marker_start = 2,
                         chr = 'chr',
                         chr_pos = 'pos')

# Check for genotype call  error for each SNP
geno_mat <- geno_error(x = proc_plate1$ordered_geno,
                  rp = 1,
                  dp = 7,
                  sep = ':')

# Convert to numeric format for plotting
num_geno <- kasp_numeric(x = geno_mat$geno_good,
                        rp_row = 1,
                        dp_row = 7)

}

}
