% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{list_columns}
\alias{list_columns}
\title{Get table schema details}
\usage{
list_columns(
  con = NULL,
  table_name = c("variants", "annotations", "genotypes", "metadata"),
  connect_db_mode = c("local", "online")
)
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{table_name}{A character value specifying the name of the table to inspect.
Must be one of `"variants"`, `"annotations"`, `"genotypes"`, or `"metadata"`.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A data frame containing the schema details of the target table.
}
\description{
Retrieves the schema (column names and data types) for a specified table.
It can operate in 'local' or 'online' mode.
}
\examples{
\donttest{
# --- Online Mode ---
# Load the package
library(panGenomeBreedr)

# Get schema for the "variants" table from the online database
online_schema <- list_columns(table_name = "variants", connect_db_mode = 'online')
print(online_schema)

# --- Offline Mode ---
# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# Inspect the schema of the local "variants" table
local_schema <- list_columns(con = con_local, table_name = "variants")
# print(local_schema)

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

}
