% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{list_sqlite_tables}
\alias{list_sqlite_tables}
\title{List all tables in the database}
\usage{
list_sqlite_tables(con)
}
\arguments{
\item{con}{A DBI connection object to the local database.}
}
\value{
A character vector of table names in the database.
}
\description{
List all tables in the database
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"
con <- connect_local_db(folder_path = my_db_folder)
list_sqlite_tables(con)
disconnect_local_db(con)
}

}
