% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{list_sqlite_tables}
\alias{list_sqlite_tables}
\title{List all tables in the database (local)}
\usage{
list_sqlite_tables(con)
}
\arguments{
\item{con}{A DBI connection object to the local database.}
}
\value{
A character vector of table names in the database.
}
\description{
List all tables in the database (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# List tables
list_sqlite_tables_result <- list_sqlite_tables(con)
print(list_sqlite_tables_result)

# Disconnect at the end of your session
disconnect_local_db(con)

}

}
