% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{list_table_columns}
\alias{list_table_columns}
\title{Get table schema details (local)}
\usage{
list_table_columns(con, table_name = c("variants", "annotations", "genotypes"))
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{table_name}{A character value specifying the name of the table to inspect.
Must be one of \code{"variants"}, \code{"annotations"}, or \code{"genotypes"}.}
}
\value{
A data frame containing the schema details of the target table, including
  column names and data types.
}
\description{
Get table schema details (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Inspect the schema layout of the variants table
column_schema <- list_table_columns(con, table_name = "variants")

# Disconnect at the end of your session
disconnect_local_db(con)
}

}
