% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{list_table_columns}
\alias{list_table_columns}
\title{Check the column names and types for any table in a SQLite database.}
\usage{
list_table_columns(
  db_path,
  table_name = c("variants", "annotations", "genotypes")
)
}
\arguments{
\item{db_path}{A character value indicating the path to the SQLite database.}

\item{table_name}{A character value specifying the name of any any table in a
SQLite database.}
}
\value{
A data frame of table column information in the SQLite database.
}
\description{
Check the column names and types for any table in a SQLite database.
}
\examples{
\donttest{
library(panGenomeBreedr)

# Define tempdir
path <- tempdir()

# Mini SQLite database
mini_db <-  system.file("extdata", "mini_sorghum_variant_vcf.db.gz",
                     package = "panGenomeBreedr",
                     mustWork = TRUE)

# Path to SQLite databases: INDEL and SNP
mini_db_path <- file.path(path, 'mini_sorghum_variant_vcf.db')

# Unzip compressed mini database and save in tempdir
R.utils::gunzip(mini_db,
               destname = mini_db_path,
               remove = FALSE)

# List column names and type of variable in the variants table in the database
list_table_columns(mini_db_path, table_name = "variants")

# Clean tempdir
contents <- list.files(tempdir(),
                             full.names = TRUE,
                             recursive = TRUE,
                             all.files = TRUE,
                             include.dirs = TRUE)
unlink(contents, recursive = TRUE, force = TRUE)
}

}
