% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_functions.R
\name{parse_marker_ns}
\alias{parse_marker_ns}
\title{Parse marker names with a common pattern containing chromosome numbers and
positions into a map file.}
\usage{
parse_marker_ns(x, sep = "_", prefix = "S")
}
\arguments{
\item{x}{A character vector containing the original marker names to be parsed.}

\item{sep}{A character value that serves as a unique separator between chromosome
positions and other components of the marker name; default value is an underscore.}

\item{prefix}{A character value that represents a common pattern in marker
names that precedes the chromome number; the default value is `S`.}
}
\value{
A data frame of map file consisting of the original marker names,
chromosome numbers and positions.
# example code
\donttest{
library(panGenomeBreedr)
snps <- paste0('S', 1:10, '_', 101:110)
map_file <- parse_marker_ns(x = snps, sep = '_', prefix = 'S')
}
}
\description{
Parse marker names with a common pattern containing chromosome numbers and
positions into a map file.
}
\details{
The marker names to be parsed into a map file must contain the chromosome numbers
and their positions
}
