% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_list_table_columns}
\alias{pg_list_table_columns}
\title{Get table schema details (online)}
\usage{
pg_list_table_columns(
  table_name = c("variants", "annotations", "genotypes", "metadata")
)
}
\arguments{
\item{table_name}{A character value specifying the name of the table.
Defaults to "variants". Available options are "variants", "annotations",
"genotypes", or "metadata".}
}
\value{
A data frame containing column metadata.
}
\description{
Get table schema details (online)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Get column metadata for the "variants" table
pg_list_table_columns_result <- pg_list_table_columns(table_name = "variants")
print(pg_list_table_columns_result)
}

}
