% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_list_tables}
\alias{pg_list_tables}
\title{List all tables in the database (online)}
\usage{
pg_list_tables()
}
\value{
A character vector of table names.
}
\description{
List all tables in the database (online)
}
\examples{
\donttest{
# Load the package
library(panGenomeBreedr)

# List all tables in the online database
pg_list_tables_result <- pg_list_tables()
print(pg_list_tables_result)
}

}
