% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{pg_list_tables}
\alias{pg_list_tables}
\title{List all tables in the PostgreSQL database}
\usage{
pg_list_tables()
}
\value{
A character vector of table names.
}
\description{
This function connects to the public panGenomeBreedr API and retrieves the names of
all tables within the database.
}
