% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_variant_discovery.R
\name{pg_list_tables}
\alias{pg_list_tables}
\title{List all tables in the PostgreSQL database}
\usage{
pg_list_tables(con)
}
\arguments{
\item{con}{A \code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}
}
\value{
A character vector of table names.
}
\description{
This function connects to the PostgreSQL server and retrieves the names of
all tables within the connected database.
}
\details{
This is a wrapper around \code{DBI::dbListTables}. For PostgreSQL, this
will typically return tables in the 'public' schema unless specified otherwise
in the connection.
}
\examples{
\dontrun{
library(panGenomeBreedr)
library(DBI)

# Establish connection
con <- dbConnect(RPostgres::Postgres(),
                 dbname = "sorghum_pangenome_db",
                 host = "localhost",
                 user = "israeltawiahtetteh")

# List tables
pg_list_tables(con)

# Disconnect
dbDisconnect(con)
}

}
