% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_summarize_tables}
\alias{pg_summarize_tables}
\title{Get table names and row counts from the database (online)}
\usage{
pg_summarize_tables()
}
\value{
A data frame with two columns: \code{table} and \code{n_rows}.
}
\description{
Get table names and row counts from the database (online)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Summarize all tables in the online database
pg_summarize_tables_result <- pg_summarize_tables()
print(pg_summarize_tables_result)
}

}
