% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{pg_summarize_tables}
\alias{pg_summarize_tables}
\title{Summarize names and row counts for each table}
\usage{
pg_summarize_tables()
}
\value{
A data frame with two columns: \code{table} and \code{n_rows}.
}
\description{
This function connects to the public panGenomeBreedr API and returns a
summary data frame containing the table names and their respective row counts.
}
