% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_variant_discovery.R
\name{pg_summarize_tables}
\alias{pg_summarize_tables}
\title{Summarize names and row counts for each table in the PostgreSQL database}
\usage{
pg_summarize_tables(con)
}
\arguments{
\item{con}{A \code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}
}
\value{
A data frame with two columns:
\itemize{
  \item \code{table}: Name of the table.
  \item \code{n_rows}: Total number of rows in that table.
}
}
\description{
This function iterates through all tables in the connected PostgreSQL database
and returns a summary data frame containing the table names and their
respective row counts.
}
\details{
For very large tables, \code{COUNT(*)} can take
a few seconds as PostgreSQL ensures transaction accuracy.
}
\examples{
\dontrun{
library(panGenomeBreedr)
library(DBI)

con <- dbConnect(RPostgres::Postgres(),
                 dbname = "sorghum_pangenome_db",
                 host = "localhost",
                 user = "israeltawiahtetteh")

# Summarize all tables
db_summary <- pg_summarize_tables(con)
print(db_summary)

dbDisconnect(con)
}

}
