% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{pg_variant_impact_summary}
\alias{pg_variant_impact_summary}
\title{Get variant impact summary}
\usage{
pg_variant_impact_summary()
}
\value{
A data frame in wide format where each row is a chromosome and
columns represent the counts for each impact category.
}
\description{
This function connects to the public panGenomeBreedr API and summarizes the
distribution of mutation impacts (e.g., HIGH, MODERATE, LOW, MODIFIER)
across chromosomes.
}
