% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{pg_variant_stats}
\alias{pg_variant_stats}
\title{Get variant statistics from the database (online)}
\usage{
pg_variant_stats(include_annotations = TRUE)
}
\arguments{
\item{include_annotations}{A logical value indicating whether to include
statistics for the annotations table. Defaults to \code{TRUE}.}
}
\value{
A data frame containing variant statistics.
}
\description{
Get variant statistics from the database (online)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Get variant statistics from the online database, including annotation counts
pg_variant_stats_result <- pg_variant_stats(include_annotations = TRUE)
print(pg_variant_stats_result)

# Get variant statistics without annotation counts
pg_variant_stats_no_ann <- pg_variant_stats(include_annotations = FALSE)
print(pg_variant_stats_no_ann)
}
}
