% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{pg_variant_stats}
\alias{pg_variant_stats}
\title{Get variant statistics from the PostgreSQL database}
\usage{
pg_variant_stats(include_annotations = TRUE)
}
\arguments{
\item{include_annotations}{A logical value indicating whether to include
statistics for the annotations table. Defaults to \code{TRUE}.}
}
\value{
A data frame containing variant statistics.
}
\description{
This function connects to the public panGenomeBreedr API and calculates
summary statistics for variants per chromosome, including variant counts
and genomic ranges.
}
