% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{plot_gene_model}
\alias{plot_gene_model}
\title{Plot the gene model using genomic coordinates (local)}
\usage{
plot_gene_model(gene_df)
}
\arguments{
\item{gene_df}{A data frame containing the genomic coordinates, expected to
be the output from \code{gene_coord_gff}.}
}
\value{
A \code{ggplot} object representing the structural gene model.
}
\description{
Visualizes the gene structure with adjusted component heights,
a cleaned legend, and a strand orientation arrow. Introns are implicitly
represented by the exposed transcript backbone.
}
\examples{
\donttest{
library(panGenomeBreedr)
# Path to GFF3 file
gff_path <- "https://raw.githubusercontent.com/awkena/panGB/main/Sbicolor_730_v5.1.gene.gff3.gz"
gene_features <- gene_coord_gff(gene_name = "Sobic.005G213600",
                                gff_path = gff_path)
# Plot gene model for Sobic.005G213600
plot_gene_model(gene_df = gene_features)
}
}
