% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ld_geodesic}
\alias{plot_ld_geodesic}
\title{Generate Geodesic Landscape Plot and Extract Haploblock Data}
\usage{
plot_ld_geodesic(
  ld_df,
  query_db_geno,
  query_db_annot,
  metric = "R2",
  threshold = 0.2,
  block_threshold = 0.8,
  show_variant_labels = TRUE,
  target_variant_ids = NULL,
  filled = FALSE,
  palette_option = "magma"
)
}
\arguments{
\item{ld_df}{A long-format data frame of pairwise LD statistics containing
columns \code{variant_1}, \code{variant_2}, and the column matching \code{metric}.}

\item{query_db_geno}{A data frame with variant genotype metadata. Must include a
\code{variant_id} column and a genomic position column (e.g., 'pos').}

\item{query_db_annot}{A data frame with variant annotation metadata. Must include
\code{variant_id} and a functional \code{impact} column.}

\item{metric}{Character string. The linkage metric to visualize ("R2" or "D_prime").}

\item{threshold}{Numeric. The minimum LD value to display. Pairs below this
value fade to white background spaces. Defaults to 0.2.}

\item{block_threshold}{Numeric. The threshold at which to group and isolate block variants. Defaults to 0.8.}

\item{show_variant_labels}{Logical. If TRUE (default), displays variant IDs above the spine.}

\item{target_variant_ids}{A character vector of variant IDs to highlight on the plot.}

\item{filled}{Logical. If `TRUE`, renders filled geodesic polygons. If `FALSE`
(default), renders line-based trajectories.}

\item{palette_option}{Character string. The `viridis` color palette to use for
mapping LD values these include: "magma", "plasma", "inferno", "viridis", "mako", "cividis", "rocket" and "turbo". Defaults to `"magma"`.}
}
\value{
A structured \code{list} containing two named assets:
\itemize{
  \item \code{plot}: A ggplot object showing the geodesic LD landscape.
  \item \code{table}: A data frame where columns represent haploblocks, listing their constituent variants and impacts.
}
}
\description{
Generate Geodesic Landscape Plot and Extract Haploblock Data
}
\examples{
\dontrun{
query_annot <- panGenomeBreedr::pg_query_db(
table_name = "annotations",
chrom = "Chr03",
gene_name = "Sobic.003G421300",
start = 79037682,
end = 79039091
)

query_geno <- panGenomeBreedr::pg_query_db(
table_name = "genotypes",
chrom = "Chr03",
gene_name = "Sobic.003G421300",
start = 79037682,
end = 79039091
)

# Compute your updated wide data containing distance tracking columns
ld_results <- compute_LD(
df = query_geno,
target_variant_ids = NULL,
genotype_start_col = 11
)

# Generate the plot and table package
result <- plot_ld_geodesic(
  ld_df = ld_results,
  query_db_geno = query_geno,
  query_db_annot = query_annot,
  metric = "R2",
  target_variant_ids = c("INDEL_Chr03_79037889", "SNP_Chr03_79037855","SNP_Chr03_79037944"),
  threshold = 0.2,
  block_threshold = 0.8
)

# Access the plot
result$plot

# Access the haploblock table
print(result$table)
}
}
