% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{plot_variant_hotspot}
\alias{plot_variant_hotspot}
\title{Plot genomic variant hotspots (SNPs and INDELs) extracted from the database (local)}
\usage{
plot_variant_hotspot(
  var_df,
  chrom_col = "chrom.x",
  pos_col = "pos.x",
  maf_col = "minor_allele_freq",
  start_pos = NULL,
  end_pos = NULL
)
}
\arguments{
\item{var_df}{A data frame containing variant data. Required columns:
\code{variant_type} (character: "SNP", "INDEL"), \code{variant_id},
\code{impact} (character: "MODIFIER", "LOW", "MODERATE", "HIGH"), and a numeric MAF column.}

\item{chrom_col}{A character string specifying the chromosome column.}

\item{pos_col}{A character string specifying the variant position column.}

\item{maf_col}{A character string specifying the minor allele frequency column.}

\item{start_pos}{A numeric integer for the region start. Default is the minimum
position in var_df.}

\item{end_pos}{A numeric integer for the region end. Default is the maximum
position in var_df.}
}
\value{
A \code{ggplot} object representing the variant hotspots.
}
\description{
Visualizes variant distributions across a genomic region.
Variants are mapped by position, with impact levels determining their
vertical placement (lollipop style), variant types mapped to shapes, and
minor allele frequencies (MAF) mapped to point size.
}
\examples{
\donttest{
library(panGenomeBreedr)
# Extract variants within the candidate gene: Sobic.005G213600
pg_gt_region <- pg_query_db(table_name = "genotypes",
                            chrom = "Chr05",
                            start = 75104537,
                            end = 75106403
                            )

#' # Extract annotations for the candidate gene
pg_annota_region <- pg_query_db(table_name = "annotations",
                                chrom = "Chr05",
                                start = 75104537,
                                end = 75106403,
                                gene_name = "Sobic.005G213600"
                                )

var_df <- merge(pg_annota_region, pg_gt_region, by = "variant_id")[,1:23]

plot_variant_hotspot(var_df)
}
}
