% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{query_ann_summary}
\alias{query_ann_summary}
\title{Get summarized functional annotations, impacts and variant types for a genomic region (local)}
\usage{
query_ann_summary(
  con,
  chrom,
  start,
  end,
  annotations_table = "annotations",
  variants_table = "variants"
)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{chrom}{A character value specifying the target chromosome name (e.g., \code{"Chr05"}).}

\item{start}{Integer. Start coordinate for the target window region.}

\item{end}{Integer. End coordinate for the target window region.}

\item{annotations_table}{A character value specifying the name of the annotations reference table. 
Defaults to \code{"annotations"}.}

\item{variants_table}{A character value specifying the name of the variants reference table. 
Defaults to \code{"variants"}.}
}
\value{
A list object containing three summarized summary data frames:
\itemize{
  \item \code{annotation_summary}: Counts of distinct functional mutations grouped by type.
  \item \code{impact_summary}: Counts of predicted functional impacts grouped by type.
  \item \code{variant_type_totals}: Entire marker breakdown counts across the defined coordinate window.
}
}
\description{
Get summarized functional annotations, impacts and variant types for a genomic region (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Extract functional distribution summaries across a specific gene locus window
query_ann_summary_result <- query_ann_summary(con = con,
                                        chrom = "Chr05",
                                        start = 75104537,
                                        end = 75106403)
print(query_ann_summary_result)


# Disconnect at the end of your session
disconnect_local_db(con)
}

}
