% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{query_by_af}
\alias{query_by_af}
\title{Query variants by allele frequency (local)}
\usage{
query_by_af(
  con,
  min_af = 0,
  max_af = 1,
  chrom = NULL,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{min_af}{Numeric. Minimum alternate allele frequency threshold for filtering. 
Must fall between 0 and 1. Defaults to \code{0}.}

\item{max_af}{Numeric. Maximum alternate allele frequency threshold for filtering. 
Must fall between 0 and 1. Defaults to \code{1}.}

\item{chrom}{A character value specifying the target chromosome name (e.g., \code{"Chr05"}).}

\item{start}{Integer. Start coordinate for the target window region.}

\item{end}{Integer. End coordinate for the target window region.}
}
\value{
A data frame of variants passing the frequency thresholds within the defined region, 
  tracking their coordinates, calculated reference allele frequencies (\code{ref_af}), 
  and alternate allele frequencies (\code{alt_af}).
}
\description{
Query variants by allele frequency (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Query a specific locus window and filter for common variants (MAF >= 5\%)
query_by_af_result <- query_by_af(con = con,
                                min_af = 0.05,
                                max_af = 0.95,
                                chrom = "Chr05",
                                start = 75104537,
                                end = 75106403)
print(query_by_af_result)

# Disconnect at the end of your session
disconnect_local_db(con)
}

}
