% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{query_genotypes}
\alias{query_genotypes}
\title{Query genotypes for one or more variant IDs from a wide-format genotype table.}
\usage{
query_genotypes(db_path, variant_ids)
}
\arguments{
\item{db_path}{The path to your SQLite or PostgreSQL database.}

\item{variant_ids}{A character vector of variant IDs to query.}
}
\value{
A data.frame of genotype data (samples x variants).
}
\description{
Query genotypes for one or more variant IDs from a wide-format genotype table.
}
\examples{
\donttest{
# example code
library(panGenomeBreedr)

# Define tempdir
path <- tempdir()

# Mini SQLite database
mini_db <-  system.file("extdata", "mini_sorghum_variant_vcf.db.gz",
                     package = "panGenomeBreedr",
                     mustWork = TRUE)

# Path to SQLite databases: INDEL and SNP
mini_db_path <- file.path(path, 'mini_sorghum_variant_vcf.db')

# Unzip compressed mini database and save in tempdir
R.utils::gunzip(mini_db,
               destname = mini_db_path,
               remove = FALSE)

# Extract all genotypes for all samples for any set of variants
geno_filter <- query_genotypes(db_path = mini_db_path,
                               variant_ids = c("INDEL_Chr05_75104541",
                                               "SNP_Chr05_75104557"))

# Clean tempdir
contents <- list.files(tempdir(),
                             full.names = TRUE,
                             recursive = TRUE,
                             all.files = TRUE,
                             include.dirs = TRUE)
unlink(contents, recursive = TRUE, force = TRUE)
}

}
