% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_validation.R
\name{scale_axis}
\alias{scale_axis}
\title{Normalize FAM and HEX fluorescence values between 0 and 1}
\usage{
scale_axis(x)
}
\arguments{
\item{x}{A numeric vector of FAM or HEX fluorescence values}
}
\value{
A numeric vector of normalized values between 0 and 1
}
\description{
Normalize FAM and HEX fluorescence values between 0 and 1
}
\examples{
# example code
library(panGenomeBreedr)
# Get Plate 1
dat1 <- panGenomeBreedr::kasp_dat[1:96,]
FAM_scaled <- scale_axis(dat1$X)

}
