% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online_variant_discovery.R
\name{set_api_url}
\alias{set_api_url}
\title{Set the API endpoint URL for your private database}
\usage{
set_api_url(url)
}
\arguments{
\item{url}{Character. The full URL to your hosted API endpoint
(e.g., "http://rice-genomics.myuniversity.edu:8000").}
}
\description{
Set the API endpoint URL for your private database
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Set a custom API endpoint URL
set_api_url("http://my-custom-api.com:8000")

}
}
