% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{summarize_sqlite_tables}
\alias{summarize_sqlite_tables}
\title{Get table names and row counts from the database (local)}
\usage{
summarize_sqlite_tables(con)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}
}
\value{
A data frame containing two columns: \code{table} (the table name) 
  and \code{n_rows} (the total number of rows in that table).
}
\description{
Queries the active local database to instantly retrieve a list of all mounted 
tables along with their respective total row counts.
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Get row counts for all mounted Parquet views
database_summary <- summarize_sqlite_tables(con)

# Disconnect at the end of your session
disconnect_local_db(con)
}

}
