% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{summarize_sqlite_tables}
\alias{summarize_sqlite_tables}
\title{Name and row count for each table in SQLite database.}
\usage{
summarize_sqlite_tables(db_path)
}
\arguments{
\item{db_path}{A character value indicating the path to the SQLite database.}
}
\value{
A data frame row counts of tables in SQLite database.
}
\description{
Name and row count for each table in SQLite database.
}
\examples{
\donttest{
library(panGenomeBreedr)

# Define tempdir
path <- tempdir()

# Mini SQLite database
mini_db <-  system.file("extdata", "mini_sorghum_variant_vcf.db.gz",
                     package = "panGenomeBreedr",
                     mustWork = TRUE)

# Path to SQLite databases: INDEL and SNP
mini_db_path <- file.path(path, 'mini_sorghum_variant_vcf.db')

# Unzip compressed mini database and save in tempdir
R.utils::gunzip(mini_db,
               destname = mini_db_path,
               remove = FALSE)

# Row counts for each table in the database
summarize_sqlite_tables(mini_db_path)

# Clean tempdir
contents <- list.files(tempdir(),
                             full.names = TRUE,
                             recursive = TRUE,
                             all.files = TRUE,
                             include.dirs = TRUE)
unlink(contents, recursive = TRUE, force = TRUE)
}

}
