% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_discovery.R
\name{summarize_variant_impacts}
\alias{summarize_variant_impacts}
\title{Get summary of database variant impacts by chromosome}
\usage{
summarize_variant_impacts(con = NULL, connect_db_mode = c("local", "online"))
}
\arguments{
\item{con}{A DBI connection object to the local database. Required only when
`connect_db_mode` is 'local'. Defaults to `NULL`.}

\item{connect_db_mode}{A character string specifying the connection mode.
Can be either `'local'` (default) or `'online'`.}
}
\value{
A wide-format data frame where the first column is \code{chrom}, followed by
  pivoted columns for each impact type (e.g., \code{impact_HIGH}, \code{impact_MODERATE}),
  containing their respective mutation counts.
}
\description{
This function summarizes the counts of different variant impact levels (e.g.,
HIGH, MODERATE, LOW) for each chromosome and returns the result in a
wide-format data frame. It can operate in 'local' or 'online' mode.
}
\examples{
\donttest{
# --- Online Mode ---
# Load the package
library(panGenomeBreedr)

# Get variant impact summary from the online database
online_impact <- summarize_variant_impacts(connect_db_mode = 'online')
print(online_impact)

# --- Offline Mode ---
# Locate the package example database folder
my_db_folder <- system.file("extdata", "pangenome_scale_db",
                           package = "panGenomeBreedr",
                           mustWork = TRUE)

# Establish a virtual connection to the offline database engine
con_local <- connect_local_db(folder_path = my_db_folder)

# Generate the wide-format impact profile matrix
local_impact <- summarize_variant_impacts(con = con_local)
# print(local_impact)

# Disconnect at the end of your session
disconnect_local_db(con_local)
}

}
