% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{variant_impact_summary}
\alias{variant_impact_summary}
\title{Get summary of database variant impacts by chromosome (local)}
\usage{
variant_impact_summary(con)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}
}
\value{
A wide-format data frame where the first column is \code{chrom}, followed by
  pivoted columns for each impact type (e.g., \code{impact_HIGH}, \code{impact_MODERATE}),
  containing their respective mutation counts.
}
\description{
Get summary of database variant impacts by chromosome (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = my_db_folder)

# Generate the wide-format impact profile matrix
impact_matrix <- variant_impact_summary(con)

# Disconnect at the end of your session
disconnect_local_db(con)

}

}
