% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offline_variant_discovery.R
\name{variant_stats}
\alias{variant_stats}
\title{Get variant statistics from the database (local)}
\usage{
variant_stats(con, include_annotations = TRUE)
}
\arguments{
\item{con}{A DBI connection object to the local database generated by \code{connect_local_db()}.}

\item{include_annotations}{A logical value indicating whether to include
statistics for the annotations table. Defaults to \code{TRUE}.}
}
\value{
A data frame of variant statistics grouped by chromosome.
}
\description{
Get variant statistics from the database (local)
}
\examples{
\dontrun{
# Load the package
library(panGenomeBreedr)

# Define the path to the curated_sorghum_variant_resource folder
my_db_folder <- "~/Desktop/curated_sorghum_variant_resource"

# Establish the connection
con <- connect_local_db(folder_path = "~/Desktop/curated_sorghum_variant_resource")

# Get variant statistics across all chromosomes
stats_df <- variant_stats(con, include_annotations = TRUE)

# Disconnect at the end of your session
disconnect_local_db(con)

}

}
