% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_variant_discovery.R
\name{variant_stats}
\alias{variant_stats}
\title{Get variants statistics stored in SQLite database}
\usage{
variant_stats(db_path, include_annotations = TRUE)
}
\arguments{
\item{db_path}{A character value indicating the path to the SQLite database.}

\item{include_annotations}{A logical value indicating whether to include
statistics for the annotations table.}
}
\value{
A data frame of variant statistics in the SQLite databases
}
\description{
Get variants statistics stored in SQLite database
}
\examples{
\donttest{
library(panGenomeBreedr)

# Define tempdir
path <- tempdir()

# Mini SQLite database
mini_db <-  system.file("extdata", "mini_sorghum_variant_vcf.db.gz",
                     package = "panGenomeBreedr",
                     mustWork = TRUE)

# Path to SQLite databases: INDEL and SNP
mini_db_path <- file.path(path, 'mini_sorghum_variant_vcf.db')

# Unzip compressed mini database and save in tempdir
R.utils::gunzip(mini_db,
               destname = mini_db_path,
               remove = FALSE)

# Get variant statistics in the SQLite database
variant_stats(mini_db_path)

# Clean tempdir
contents <- list.files(tempdir(),
                             full.names = TRUE,
                             recursive = TRUE,
                             all.files = TRUE,
                             include.dirs = TRUE)
unlink(contents, recursive = TRUE, force = TRUE)
}

}
